+++ sentry/src/main/java/io/sentry/SentryOptions.java
@@ -454,8 +454,8 @@ public class SentryOptions {
 
   @ApiStatus.Experimental private boolean enableBackpressureHandling = false;
 
-  /** Whether to enable startup profiling, depending on profilesSampler or profilesSampleRate. */
-  private boolean enableStartupProfiling = false;
+  /** Whether to profile app launches, depending on profilesSampler or profilesSampleRate. */
+  private boolean enableAppStartProfiling = false;
 
   /**
    * Profiling traces rate. 101 hz means 101 traces in 1 second. Defaults to 101 to avoid possible
@@ -2154,22 +2154,22 @@ public class SentryOptions {
   }
 
   /**
-   * Whether to enable startup profiling, depending on profilesSampler or profilesSampleRate.
-   * Depends on {@link SentryOptions#isProfilingEnabled()}
+   * Whether to profile app launches, depending on profilesSampler or profilesSampleRate. Depends on
+   * {@link SentryOptions#isProfilingEnabled()}
    *
-   * @return true if startup profiling should be started.
+   * @return true if app launches should be profiled.
    */
-  public boolean isEnableStartupProfiling() {
-    return isProfilingEnabled() && enableStartupProfiling;
+  public boolean isEnableAppStartProfiling() {
+    return isProfilingEnabled() && enableAppStartProfiling;
   }
 
   /**
-   * Whether to enable startup profiling, depending on profilesSampler or profilesSampleRate.
+   * Whether to profile app launches, depending on profilesSampler or profilesSampleRate.
    *
-   * @param enableStartupProfiling true if startup profiling should be started.
+   * @param enableAppStartProfiling true if app launches should be profiled.
    */
-  public void setEnableStartupProfiling(boolean enableStartupProfiling) {
-    this.enableStartupProfiling = enableStartupProfiling;
+  public void setEnableAppStartProfiling(boolean enableAppStartProfiling) {
+    this.enableAppStartProfiling = enableAppStartProfiling;
   }
 
   /**